Saving computed P-POSSUM fields

+++ database/migrations/2018_03_08_205848_create_patients_table.php
@@ -82,8 +82,8 @@ class CreatePatientsTable extends Migration
             $table->integer('ppurgency')->nullable();
             $table->integer('ppphysiology')->nullable();
             $table->integer('ppopscore')->nullable();
-            $table->integer('ppmorbidity')->nullable();
-            $table->integer('ppmortality')->nullable();
+            $table->float('ppmorbidity', 3, 1)->nullable();
+            $table->float('ppmortality', 3, 1)->nullable();
             $table->timestamps();
         });
     }
